Reorganize patch and extension package structure
