+++ extensions/music/src/main/java/app/morphe/extension/music/patches/scrobbling/listenbrainz/ListenBrainzTokenStore.java
@@ -5,7 +5,7 @@
  * See the included NOTICE file for GPLv3 §7(b) and §7(c) terms that apply to this code.
  */
 
-package app.morphe.extension.music.patches.listenbrainz;
+package app.morphe.extension.music.patches.scrobbling.listenbrainz;
 
 import app.morphe.extension.music.settings.Settings;
